Add ability to override validators via loom.yml (#435)
The validator override is implemented in the gamechain-validator-override branch of loomchain, that branch is based off the loomchain & go-loom revisions that the last released gamechain build (413) was built from. The gamechain master branch is now locked to the gamechain-validator-override branch of loomchain so changes there don't break future gamechain builds, to intentionally upgrade gamechain to a newer loomchain version update LOOMCHAIN_GIT_REV to a new revision/branch/tag.